+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [1.12.0] - 2024-06-14
+# [1.12.0] - 2024-06-20
 - Added support of Python 3.12
+- Added possibility to configure `FunctionResponseTypes` for lambda functions
 - Fixed an issue related to ARNs resolving in case of empty resource name
 - Fixed an issue related to improper filtering of resources in case of different types of filter usage
 - Fixed an error related to SQS FIFO Queue availability regions